Move load COLUMN_NUMBER_HASH logic from ckb-shared to ckb-store
